requirements.txt
psutil
allure-pytest==2.8.18
allure-python-commons==2.8.18
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.5.1
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-pylint==0.17.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
webdriver-manager==3.2.2
wrapt==1.12.1
zipp==3.3.1
xmltodict==0.12.0

